update patch according to comments left by A.Nikulin
