+++ docs/snippets_generation/swagger_ui_curlify.py
@@ -5,14 +5,15 @@ Original implementation: https://github.com/swagger-api/swagger-ui/blob/master/s
 from docs.snippets_generation import body_generator
 
 
-def generate_sample(op_spec, data_params_are_present, model_name, full_spec, base_headers, jinja_env):
+def generate_sample(op_spec, data_params_are_present, model_name, full_spec, jinja_env):
     body = None
     template_name = "snippet_curl.j2"
-    headers = dict(base_headers)
-    # add mandatory token
-    headers["Authorization"] = 'Bearer ${ACCESS_TOKEN}'
-    headers.pop('User-Agent', None)
+    headers = {
+        "Accept": "application/json",
+        "Authorization": "Bearer ${ACCESS_TOKEN}"
+    }
     if data_params_are_present:
+        headers["Content-Type"] = "application/json"
         body = body_generator.generate_model_sample(model_name, full_spec)
 
     template = jinja_env.get_template(template_name)
